Add fixes for merging with the upstream
Fix Android build:

* Define __STDC_FORMAT_MACROS if needed
* Add qalgo/base64.c, qalgo/md5.c to game module source files

Fix MSVC build:
* Avoid a zero-sized array as a member of WorldState
* Fix excessive variable shadowning warnings
* Fix some sign/size integer mismatch occurrences
* Fix some "forcing value to bool" warnings
* Include <math.h> and define __USE_MATH_DEFINES if needed
* Fix not tested yet on this platform varargs functions missing ...
* IMPORTANT: Fix shifts of a 32-bit integer on more than 32 bits
* IMPORTANT: Fix missing setSatisfyOp methods
  in script world state var classes declarations

+++ source/game/ai/ai_ground_trace_cache.cpp
@@ -35,7 +35,7 @@ void AiGroundTraceCache::GetGroundTrace( const edict_s *ent, float depth, trace_
 	edict_t *entRef = const_cast<edict_t *>( ent );
 	CachedTrace *cachedTrace = (CachedTrace *)data + ENTNUM( entRef );
 
-	if( cachedTrace->computedAt + maxMillisAgo >= level.time ) {
+	if( (int64_t)( cachedTrace->computedAt + maxMillisAgo ) >= level.time ) {
 		if( cachedTrace->depth >= depth ) {
 			trace->startsolid = cachedTrace->trace.startsolid;
 			if( cachedTrace->trace.fraction == 1.0f ) {
@@ -71,7 +71,7 @@ bool AiGroundTraceCache::TryDropToFloor( const struct edict_s *ent, float depth,
 
 	VectorCopy( ent->s.origin, result );
 
-	if( cachedTrace->computedAt + maxMillisAgo >= level.time ) {
+	if( (int64_t)( cachedTrace->computedAt + maxMillisAgo ) >= level.time ) {
 		if( cachedTrace->depth >= depth ) {
 			if( cachedTrace->trace.fraction == 1.0f ) {
 				return false;
